Renamed debug module to phony
